+++ app/client/src/ce/navigation/FocusSetters.ts
@@ -11,9 +11,9 @@ import {
   saasEditorApiIdURL,
   widgetListURL,
 } from "@appsmith/RouteBuilder";
-import { PluginType } from "../entities/Action";
+import { PluginType } from "entities/Action";
 import type { QueryListState } from "./FocusSelectors";
-import { EditorEntityTab } from "entities/IDE/constants";
+import { EditorEntityTab } from "@appsmith/entities/IDE/constants";
 
 export function setSelectedDatasource(id: string | undefined) {
   if (id) {
