+++ app/client/src/ce/navigation/FocusStrategy/NoIDEFocusStrategy.ts
@@ -1,0 +1,15 @@
+import type { FocusStrategy } from "sagas/FocusRetentionSaga";
+import NoIDEFocusElements from "../FocusElements/NoIDE";
+export const NoIDEFocusStrategy: FocusStrategy = {
+  focusElements: NoIDEFocusElements,
+  *getEntitiesForSet() {
+    return [];
+  },
+  *getEntitiesForStore() {
+    return [];
+  },
+  getEntityParentUrl(): string {
+    return "";
+  },
+  *waitForPathLoad() {},
+};
