+++ app/client/src/ee/entities/IDE/constants.ts
@@ -1,0 +1,1 @@
+export * from "ce/entities/IDE/constants";
