+++ app/client/src/ee/entities/IDE/utils.ts
@@ -1,0 +1,1 @@
+export * from "ce/entities/IDE/utils";
