+++ app/client/src/ee/navigation/FocusSelectors.ts
@@ -1,0 +1,1 @@
+export * from "ce/navigation/FocusSelectors";
