+++ app/client/src/ee/navigation/FocusSetters.ts
@@ -1,0 +1,1 @@
+export * from "ce/navigation/FocusSetters";
