+++ app/client/src/ee/navigation/FocusStrategy/index.ts
@@ -1,0 +1,1 @@
+export * from "ce/navigation/FocusStrategy";
