+++ app/client/src/ee/sagas/NavigationSagas.ts
@@ -3,13 +3,13 @@ export * from "ce/sagas/NavigationSagas";
 import { ReduxActionTypes } from "ce/constants/ReduxActionConstants";
 import { all, takeEvery } from "redux-saga/effects";
 import { handleRouteChange } from "ce/sagas/NavigationSagas";
-import { handleRemoveFocusHistory } from "sagas/ContextSwitchingSaga";
+import FocusRetention from "sagas/FocusRetentionSaga";
 export default function* rootSaga() {
   yield all([
     takeEvery(ReduxActionTypes.ROUTE_CHANGED, handleRouteChange),
     takeEvery(
       ReduxActionTypes.REMOVE_FOCUS_HISTORY_REQUEST,
-      handleRemoveFocusHistory,
+      FocusRetention.handleRemoveFocusHistory.bind(FocusRetention),
     ),
     // EE sagas called after this
   ]);
