+++ app/client/src/layoutSystems/common/dropTarget/DropTargetComponent.tsx
@@ -50,7 +50,7 @@ import StarterBuildingBlocks from "./starterBuildingBlocks";
 import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
 import { useCurrentAppState } from "pages/Editor/IDE/hooks";
-import { EditorState as IDEAppState } from "entities/IDE/constants";
+import { EditorState as IDEAppState } from "@appsmith/entities/IDE/constants";
 import { isAirgapped } from "@appsmith/utils/airgapHelpers";
 
 export type DropTargetComponentProps = PropsWithChildren<{
