+++ app/client/src/navigation/FocusEntity.ts
@@ -1,29 +1,16 @@
+import type { match } from "react-router";
 import { matchPath } from "react-router";
-import {
-  API_EDITOR_ID_ADD_PATH,
-  API_EDITOR_ID_PATH,
-  BUILDER_CUSTOM_PATH,
-  BUILDER_PATH,
-  BUILDER_PATH_DEPRECATED,
-  CURL_IMPORT_PAGE_PATH,
-  DATA_SOURCES_EDITOR_ID_PATH,
-  INTEGRATION_EDITOR_PATH,
-  JS_COLLECTION_ID_PATH,
-  ADD_PATH,
-  QUERIES_EDITOR_ID_ADD_PATH,
-  QUERIES_EDITOR_ID_PATH,
-  WIDGETS_EDITOR_ID_PATH,
-} from "constants/routes";
-import {
-  SAAS_EDITOR_API_ID_ADD_PATH,
-  SAAS_EDITOR_API_ID_PATH,
-  SAAS_EDITOR_DATASOURCE_ID_PATH,
-} from "pages/Editor/SaaSEditor/constants";
+import { ADD_PATH, CURL_IMPORT_PAGE_PATH } from "constants/routes";
 import { TEMP_DATASOURCE_ID } from "constants/Datasource";
-import { EditorState } from "../entities/IDE/constants";
+import type { IDEType } from "@appsmith/entities/IDE/constants";
+import { EditorState, EntityPaths } from "@appsmith/entities/IDE/constants";
+import {
+  getBaseUrlsForIDEType,
+  getIDETypeByUrl,
+} from "@appsmith/entities/IDE/utils";
+import { memoize } from "lodash";
 
 export enum FocusEntity {
-  PAGE = "PAGE",
   API = "API",
   CANVAS = "CANVAS",
   DATASOURCE_LIST = "DATASOURCE_LIST",
@@ -55,75 +42,54 @@ export interface FocusEntityInfo {
   entity: FocusEntity;
   id: string;
   appState: EditorState;
+  params: MatchEntityFromPath;
+}
+
+const getMatchPaths = memoize((type: IDEType): string[] => {
+  const basePaths = getBaseUrlsForIDEType(type);
+  return EntityPaths.reduce((previousValue, currentValue) => {
+    const toAdd = basePaths.map((b) => b + currentValue);
+    return previousValue.concat(...toAdd);
+  }, [] as string[]).concat(basePaths);
+});
+
+export interface MatchEntityFromPath {
+  applicationId?: string;
+  customSlug?: string;
+  applicationSlug?: string;
+  packageId?: string;
+  moduleId?: string;
+  workflowId?: string;
+  pageSlug?: string;
+  apiId?: string;
+  datasourceId?: string;
+  pluginPackageName?: string;
+  queryId?: string;
+  appId?: string;
   pageId?: string;
+  collectionId?: string;
+  widgetIds?: string;
+  selectedTab?: string;
+  entity?: string;
+}
+
+function matchEntityFromPath(path: string): match<MatchEntityFromPath> | null {
+  const ideType = getIDETypeByUrl(path);
+  const matchPaths = getMatchPaths(ideType);
+  return matchPath(path, {
+    path: matchPaths,
+    exact: true,
+  });
 }
 
 export function identifyEntityFromPath(path: string): FocusEntityInfo {
-  const match = matchPath<{
-    apiId?: string;
-    datasourceId?: string;
-    pluginPackageName?: string;
-    queryId?: string;
-    appId?: string;
-    pageId?: string;
-    collectionId?: string;
-    widgetIds?: string;
-    selectedTab?: string; // Datasource creation/list screen
-    entity?: string;
-  }>(path, {
-    path: [
-      BUILDER_PATH_DEPRECATED + API_EDITOR_ID_ADD_PATH,
-      BUILDER_PATH_DEPRECATED + API_EDITOR_ID_PATH,
-      BUILDER_PATH + API_EDITOR_ID_ADD_PATH,
-      BUILDER_PATH + API_EDITOR_ID_PATH,
-      BUILDER_CUSTOM_PATH + API_EDITOR_ID_ADD_PATH,
-      BUILDER_CUSTOM_PATH + API_EDITOR_ID_PATH,
-      BUILDER_PATH_DEPRECATED + QUERIES_EDITOR_ID_ADD_PATH,
-      BUILDER_PATH_DEPRECATED + QUERIES_EDITOR_ID_PATH,
-      BUILDER_PATH + QUERIES_EDITOR_ID_ADD_PATH,
-      BUILDER_PATH + QUERIES_EDITOR_ID_PATH,
-      BUILDER_CUSTOM_PATH + QUERIES_EDITOR_ID_ADD_PATH,
-      BUILDER_CUSTOM_PATH + QUERIES_EDITOR_ID_PATH,
-      BUILDER_PATH_DEPRECATED + DATA_SOURCES_EDITOR_ID_PATH,
-      BUILDER_PATH + DATA_SOURCES_EDITOR_ID_PATH,
-      BUILDER_CUSTOM_PATH + DATA_SOURCES_EDITOR_ID_PATH,
-      BUILDER_PATH_DEPRECATED + INTEGRATION_EDITOR_PATH,
-      BUILDER_PATH + INTEGRATION_EDITOR_PATH,
-      BUILDER_CUSTOM_PATH + INTEGRATION_EDITOR_PATH,
-      BUILDER_PATH + SAAS_EDITOR_DATASOURCE_ID_PATH,
-      BUILDER_CUSTOM_PATH + SAAS_EDITOR_DATASOURCE_ID_PATH,
-      BUILDER_PATH_DEPRECATED + SAAS_EDITOR_API_ID_ADD_PATH,
-      BUILDER_PATH_DEPRECATED + SAAS_EDITOR_API_ID_PATH,
-      BUILDER_PATH + SAAS_EDITOR_API_ID_ADD_PATH,
-      BUILDER_PATH + SAAS_EDITOR_API_ID_PATH,
-      BUILDER_CUSTOM_PATH + SAAS_EDITOR_API_ID_ADD_PATH,
-      BUILDER_CUSTOM_PATH + SAAS_EDITOR_API_ID_PATH,
-      BUILDER_PATH_DEPRECATED + JS_COLLECTION_ID_PATH,
-      BUILDER_PATH + JS_COLLECTION_ID_PATH,
-      BUILDER_CUSTOM_PATH + JS_COLLECTION_ID_PATH,
-      BUILDER_PATH + WIDGETS_EDITOR_ID_PATH,
-      BUILDER_CUSTOM_PATH + WIDGETS_EDITOR_ID_PATH,
-      BUILDER_PATH_DEPRECATED + WIDGETS_EDITOR_ID_PATH,
-      BUILDER_PATH + WIDGETS_EDITOR_ID_PATH + ADD_PATH,
-      BUILDER_CUSTOM_PATH + WIDGETS_EDITOR_ID_PATH + ADD_PATH,
-      BUILDER_PATH_DEPRECATED + WIDGETS_EDITOR_ID_PATH + ADD_PATH,
-      BUILDER_PATH + CURL_IMPORT_PAGE_PATH,
-      BUILDER_PATH + CURL_IMPORT_PAGE_PATH + ADD_PATH,
-      BUILDER_PATH + "/:entity",
-      BUILDER_CUSTOM_PATH + "/:entity",
-      BUILDER_PATH_DEPRECATED + "/:entity",
-      BUILDER_PATH_DEPRECATED,
-      BUILDER_PATH,
-      BUILDER_CUSTOM_PATH,
-    ],
-    exact: true,
-  });
+  const match = matchEntityFromPath(path);
   if (!match) {
     return {
       entity: FocusEntity.NONE,
       id: "",
-      pageId: "",
       appState: EditorState.EDITOR,
+      params: {},
     };
   }
   if (match.params.apiId) {
@@ -131,15 +97,15 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
       return {
         entity: FocusEntity.QUERY,
         id: match.params.apiId,
-        pageId: match.params.pageId,
         appState: EditorState.EDITOR,
+        params: match.params,
       };
     }
     return {
       entity: FocusEntity.QUERY,
       id: match.params.apiId,
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.datasourceId) {
@@ -147,15 +113,15 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
       return {
         entity: FocusEntity.NONE,
         id: match.params.datasourceId,
-        pageId: match.params.pageId,
         appState: EditorState.DATA,
+        params: match.params,
       };
     } else {
       return {
         entity: FocusEntity.DATASOURCE,
         id: match.params.datasourceId,
-        pageId: match.params.pageId,
         appState: EditorState.DATA,
+        params: match.params,
       };
     }
   }
@@ -163,16 +129,16 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
     return {
       entity: FocusEntity.DATASOURCE_CREATE,
       id: match.params.selectedTab,
-      pageId: match.params.pageId,
       appState: EditorState.DATA,
+      params: match.params,
     };
   }
   if (match.params.entity === "datasource") {
     return {
       entity: FocusEntity.DATASOURCE_LIST,
       id: "",
-      pageId: match.params.pageId,
       appState: EditorState.DATA,
+      params: match.params,
     };
   }
   if (match.params.queryId) {
@@ -180,55 +146,55 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
       return {
         entity: FocusEntity.QUERY_ADD,
         id: "",
-        pageId: match.params.pageId,
         appState: EditorState.EDITOR,
+        params: match.params,
       };
     }
     return {
       entity: FocusEntity.QUERY,
       id: match.params.queryId,
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.collectionId) {
     return {
       entity: FocusEntity.JS_OBJECT,
       id: match.params.collectionId,
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.widgetIds) {
     return {
       entity: FocusEntity.PROPERTY_PANE,
       id: match.params.widgetIds,
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.entity === "widgets") {
     return {
       entity: FocusEntity.WIDGET_LIST,
       id: "",
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.entity === "queries") {
     return {
       entity: FocusEntity.QUERY_LIST,
       id: "",
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.entity === "jsObjects") {
     return {
       entity: FocusEntity.JS_OBJECT_LIST,
       id: "",
-      pageId: match.params.pageId,
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   if (match.params.entity) {
@@ -237,7 +203,7 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
         entity: FocusEntity.LIBRARY,
         id: "",
         appState: EditorState.LIBRARIES,
-        pageId: match.params.pageId,
+        params: match.params,
       };
     }
     if (match.params.entity === "settings") {
@@ -245,7 +211,7 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
         entity: FocusEntity.SETTINGS,
         id: "",
         appState: EditorState.SETTINGS,
-        pageId: match.params.pageId,
+        params: match.params,
       };
     }
   }
@@ -255,15 +221,15 @@ export function identifyEntityFromPath(path: string): FocusEntityInfo {
   ) {
     return {
       entity: FocusEntity.QUERY,
-      id: "",
-      pageId: match.params.pageId,
+      id: "curl",
       appState: EditorState.EDITOR,
+      params: match.params,
     };
   }
   return {
     entity: FocusEntity.CANVAS,
     id: "",
-    pageId: match.params.pageId,
     appState: EditorState.EDITOR,
+    params: match.params,
   };
 }
