+++ app/client/src/reducers/uiReducers/ideReducer.ts
@@ -1,7 +1,7 @@
 import { createReducer } from "utils/ReducerUtils";
 import type { ReduxAction } from "@appsmith/constants/ReduxActionConstants";
 import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
-import { EditorViewMode } from "entities/IDE/constants";
+import { EditorViewMode } from "@appsmith/entities/IDE/constants";
 
 const initialState: IDEState = {
   view: EditorViewMode.FullScreen,
